Code quality fixes
